requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bz2file==0.98
certifi==2019.6.16
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
entrypoints==0.3
Fiona==1.8.4
Flask==1.1.1
gast==0.2.2
GDAL==2.3.3
geopandas==0.4.1
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.3.4
mapclassify==2.0.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
nibabel==2.4.1
notebook==5.7.8
numpy==1.16.4
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2==2.7.6.1
ptyprocess==0.6.0
pydicom==1.3.0
Pygments==2.4.2
pyparsing==2.4.0
pyproj==1.9.6
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.0
qtconsole==4.5.1
requests==2.22.0
Rtree==0.8.3
scikit-learn==0.21.2
scipy==1.2.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
